hw2/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

itmo-latex-generator==0.2
